myproject/requirements.txt
django==5.0.2
numpy==1.24.3
pillow==10.2.0
h5py==3.10.0
keras==2.15.0 